+++ app/views/events/new.html.slim
@@ -8,28 +8,28 @@ section.event-day.overflow-visible
       div class="flex-[0_1_30%] flex justify-center"
         svg viewBox="0 0 320 512": path d="M48.1 240c-.1 2.7-.1 5.3-.1 8l0 16c0 2.7 0 5.3 .1 8L32 272c-17.7 0-32 14.3-32 32s14.3 32 32 32l28.3 0C89.9 419.9 170 480 264 480l24 0c17.7 0 32-14.3 32-32s-14.3-32-32-32l-24 0c-57.9 0-108.2-32.4-133.9-80L256 336c17.7 0 32-14.3 32-32s-14.3-32-32-32l-143.8 0c-.1-2.6-.2-5.3-.2-8l0-16c0-2.7 .1-5.4 .2-8L256 240c17.7 0 32-14.3 32-32s-14.3-32-32-32l-125.9 0c25.7-47.6 76-80 133.9-80l24 0c17.7 0 32-14.3 32-32s-14.3-32-32-32l-24 0C170 32 89.9 92.1 60.3 176L32 176c-17.7 0-32 14.3-32 32s14.3 32 32 32l16.1 0z"
 
-  = form_for @events, url: events_path, data: {controller: "new-event-form"} do |form|
-    ul.events-list data-new-event-form-target="eventsList"
+  = form_for @events, url: events_path, data: {controller: "event-fields"} do |form|
+    ul.events-list data-event-fields-target="eventsList"
       - @events.each_with_index do |event, index|
-        li data-new-event-form-target="eventFields" data-controller="remove-event-form"
-          = fields_for "events[]", event do |event_form|
+        li data-event-fields-target="eventFields" data-controller="events-form"
+          = fields_for "events[]", event do |event_fields|
             article.event-card.bg-card-bg.p-2.items-center
               div.btn.w-fit.p-2.font-quicksand.rounded-lg.bg-fgcolor.text-bgcolor
-                /= event_form.hidden_field :date
+                /= event_fields.hidden_field :date
                 div.flex.gap-2.items-center
                   svg.w-4.fill-bgcolor viewBox="0 0 448 512": path d="M128 0c17.7 0 32 14.3 32 32l0 32 128 0 0-32c0-17.7 14.3-32 32-32s32 14.3 32 32l0 32 48 0c26.5 0 48 21.5 48 48l0 48L0 160l0-48C0 85.5 21.5 64 48 64l48 0 0-32c0-17.7 14.3-32 32-32zM0 192l448 0 0 272c0 26.5-21.5 48-48 48L48 512c-26.5 0-48-21.5-48-48L0 192zm64 80l0 32c0 8.8 7.2 16 16 16l32 0c8.8 0 16-7.2 16-16l0-32c0-8.8-7.2-16-16-16l-32 0c-8.8 0-16 7.2-16 16zm128 0l0 32c0 8.8 7.2 16 16 16l32 0c8.8 0 16-7.2 16-16l0-32c0-8.8-7.2-16-16-16l-32 0c-8.8 0-16 7.2-16 16zm144-16c-8.8 0-16 7.2-16 16l0 32c0 8.8 7.2 16 16 16l32 0c8.8 0 16-7.2 16-16l0-32c0-8.8-7.2-16-16-16l-32 0zM64 400l0 32c0 8.8 7.2 16 16 16l32 0c8.8 0 16-7.2 16-16l0-32c0-8.8-7.2-16-16-16l-32 0c-8.8 0-16 7.2-16 16zm144-16c-8.8 0-16 7.2-16 16l0 32c0 8.8 7.2 16 16 16l32 0c8.8 0 16-7.2 16-16l0-32c0-8.8-7.2-16-16-16l-32 0zm112 16l0 32c0 8.8 7.2 16 16 16l32 0c8.8 0 16-7.2 16-16l0-32c0-8.8-7.2-16-16-16l-32 0c-8.8 0-16 7.2-16 16z"
                   span.text-lg Date
                 div.flex.gap-2.items-center
                   svg.w-4.fill-bgcolor viewBox="0 0 512 512": path d="M464 256A208 208 0 1 1 48 256a208 208 0 1 1 416 0zM0 256a256 256 0 1 0 512 0A256 256 0 1 0 0 256zM232 120l0 136c0 8 4 15.5 10.7 20l96 64c11 7.4 25.9 4.4 33.3-6.7s4.4-25.9-6.7-33.3L280 243.2 280 120c0-13.3-10.7-24-24-24s-24 10.7-24 24z"
                   span.text-lg Heure
-                /= event_form.datetime_field :date
+                /= event_fields.datetime_field :date
 
               div class="flex flex-col gap-1 md:flex-row md:flex-[2_0] md:justify-between md:px-1 md:py-3"
                 div.name-desc-ctn.flex.flex-col.gap-2
-                  = event_form.text_field :name,
+                  = event_fields.text_field :name,
                     class: "name w-full",
                     placeholder: "Nom de l'artiste ou de l'événement"
-                  = event_form.text_field :description,
+                  = event_fields.text_field :description,
                     class: "description w-full",
                     placeholder: "(Optionnel) Courte description, infos complémentaires..."
 
@@ -39,11 +39,11 @@ section.event-day.overflow-visible
                 div.md-tarif.flex-col.justify-between
                   svg.trash-btn.h-4.fill-fgcolor-faded.self-end.hover:fill-fgcolor.hover:cursor-pointer[
                     class="#{"hidden" if index.zero?}"
-                    data-action="click->remove-event-form#onTrashClick"
+                    data-action="click->events-form#onTrashClick"
                     viewBox="0 0 448 512"
                   ]
                     path d="M135.2 17.7L128 32 32 32C14.3 32 0 46.3 0 64S14.3 96 32 96l384 0c17.7 0 32-14.3 32-32s-14.3-32-32-32l-96 0-7.2-14.3C307.4 6.8 296.3 0 284.2 0L163.8 0c-12.1 0-23.2 6.8-28.6 17.7zM416 128L32 128 53.2 467c1.6 25.3 22.6 45 47.9 45l245.8 0c25.3 0 46.3-19.7 47.9-45L416 128z"
-                  = event_form.text_field :tarif,
+                  = event_fields.text_field :tarif,
                     class: "w-full",
                     placeholder: "Tarif (optionnel)"
           div.errors
@@ -55,7 +55,7 @@ section.event-day.overflow-visible
     /"plus" button
     div.flex.justify-center
       svg.h-10.fill-fgcolor.my-5.btn.p-0.rounded-full[
-        data-action="click->new-event-form#onPlusBtnClick"
+        data-action="click->event-fields#onPlusBtnClick"
         viewBox="0 0 512 512"
       ]
         path d="M256 512A256 256 0 1 0 256 0a256 256 0 1 0 0 512zM232 344l0-64-64 0c-13.3 0-24-10.7-24-24s10.7-24 24-24l64 0 0-64c0-13.3 10.7-24 24-24s24 10.7 24 24l0 64 64 0c13.3 0 24 10.7 24 24s-10.7 24-24 24l-64 0 0 64c0 13.3-10.7 24-24 24s-24-10.7-24-24z"
